Refactor: Move ClampedPercentage type to common folder (#2786)
## Description:

• Moved ClampedPercentage from `fuel-gas-price-algorithm` to
`crates/types/src/` for better modularity and reuse.
• Updated references in `fuel_gas_price_provider` and `genesis`
adapters.
• No functional changes, just refactoring.

## Testing:

 Ran `cargo test --all-targets`, and all tests passed.
 
 ## Linked Issue:
• ref #2773

Let me know if any further adjustments are needed!

+++ tests/tests/coins.rs
@@ -282,10 +282,7 @@ mod coin {
             .client
             .coins_to_spend(
                 &owner,
-                vec![
-                    (asset_id_a, 300, Some(MAX as u16)),
-                    (asset_id_b, 300, Some(MAX as u16)),
-                ],
+                vec![(asset_id_a, 300, Some(MAX)), (asset_id_b, 300, Some(MAX))],
                 None,
             )
             .await;
@@ -495,7 +492,7 @@ mod message_coin {
         // not enough inputs
         let coins_per_asset = context
             .client
-            .coins_to_spend(&owner, vec![(base_asset_id, 300, Some(MAX as u16))], None)
+            .coins_to_spend(&owner, vec![(base_asset_id, 300, Some(MAX))], None)
             .await;
         assert!(coins_per_asset.is_err());
         assert_eq!(
@@ -719,10 +716,7 @@ mod all_coins {
             .client
             .coins_to_spend(
                 &owner,
-                vec![
-                    (asset_id_a, 300, Some(MAX as u16)),
-                    (asset_id_b, 300, Some(MAX as u16)),
-                ],
+                vec![(asset_id_a, 300, Some(MAX)), (asset_id_b, 300, Some(MAX))],
                 None,
             )
             .await;
